backend/requirements.txt
fastapi
uvicorn
motor
pymongo
groq
python-dotenv
pydantic
sentence-transformers
numpy